RN-664: Moved all the change pulling related logic into routes/pullChanges
