+++ packages/meditrak-app-server/src/routes/sync/index.ts
@@ -3,6 +3,10 @@
  * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
  */
 
-export { CountChangesRequest, CountChangesRoute } from './CountChangesRoute';
-export { PullChangesRequest, PullChangesRoute } from './PullChangesRoute';
+export {
+  CountChangesRequest,
+  CountChangesRoute,
+  PullChangesRequest,
+  PullChangesRoute,
+} from './pullChanges';
 export { PushChangesRequest, PushChangesRoute } from './PushChangesRoute';
