+++ packages/meditrak-app-server/src/routes/sync/pullChanges/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { CountChangesRequest, CountChangesRoute } from './CountChangesRoute';
+export { PullChangesRequest, PullChangesRoute } from './PullChangesRoute';
